requirements.txt
will be using 

flask(latest version)
request (any version)